+++ com.devsquad10.company/src/main/java/com/devsquad10/company/domain/repository/CompanyRepository.java
@@ -3,8 +3,6 @@ package com.devsquad10.company.domain.repository;
 import java.util.Optional;
 import java.util.UUID;
 
-import org.springframework.data.domain.Page;
-
 import com.devsquad10.company.domain.model.Company;
 
 public interface CompanyRepository {
@@ -13,5 +11,4 @@ public interface CompanyRepository {
 
 	Company save(Company company);
 
-	Page<Company> findAll(String q, String category, int page, int size, String sort, String order);
 }
